requirements.txt
Flask==0.11.1
Flask-Mako==0.4
PyYAML==3.12
requests==2.12.4
